+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## UNRELEASED
 * Make the (Google) Analytics block configurable to provide for a new version and also other options
+* Add html class name to resource overview
+* Clear cached session openstad user when making api call or logging out
+
+## v0.10.1 (2020-12-09)
+* Cache openstad user longer when logged in to cms
 
 ## v0.10.0 (2020-12-09)
 * Make description textarea in icon-section